ok before bugs

+++ classes/adduser_spacket.cpp
@@ -10,7 +10,7 @@ void AddUser_SPacket::deserialize(QByteArray buffer)
     QDataStream in(&buffer,QIODevice::ReadOnly);
     in.setVersion(QDataStream::Qt_4_0);
 
-    short headerMe;
+    qint64 headerMe;
 
     in >> headerMe >> data;
