+++ classes/checkverifysafepacket.cpp
@@ -11,7 +11,7 @@ void CheckVerifySafePacket::deserialize(QByteArray buffer)
     QDataStream in(&buffer,QIODevice::ReadOnly);
     in.setVersion(QDataStream::Qt_4_0);
 
-    short headerMe;
+    qint64 headerMe;
 
     in >> headerMe >> Link << Answer;
