+++ classes/loginpacket.cpp
@@ -22,7 +22,7 @@ void loginPacket::deserialize(QByteArray buffer)
     QDataStream in(&buffer,QIODevice::ReadOnly);
     in.setVersion(QDataStream::Qt_4_0);
 
-    short headerMe;
+    qint64 headerMe;
 
     in >> headerMe >> JsonLoginData;
